make SIMFractureDynamics a private template
